Move buildModuleGraph to separate file
