+++ code/builders/builder-vite/src/index.ts
@@ -3,19 +3,14 @@ import { readFile } from 'node:fs/promises';
 import { fileURLToPath } from 'node:url';
 
 import { NoStatsForViteDevError } from 'storybook/internal/server-errors';
-import type {
-  Builder,
-  Middleware,
-  ModuleGraph,
-  ModuleNode,
-  Options,
-} from 'storybook/internal/types';
+import type { Builder, Middleware, ModuleGraph, Options } from 'storybook/internal/types';
 
-import type { ViteDevServer, ModuleNode as ViteModuleNode } from 'vite';
+import type { ViteDevServer } from 'vite';
 
 import { build as viteBuild } from './build';
 import type { ViteBuilder } from './types';
 import { createViteServer } from './vite-server';
+import { buildModuleGraph } from './utils/build-module-graph';
 
 export { withoutVitePlugins } from './utils/without-vite-plugins';
 export { hasVitePlugins } from './utils/has-vite-plugins';
@@ -43,69 +38,6 @@ const listeners = new Set<(moduleGraph: ModuleGraph) => void>();
 let debounce: ReturnType<typeof setTimeout> | undefined;
 let watcherChangeHandler: (() => void) | undefined;
 let waitForModuleGraph: ReturnType<typeof setInterval> | undefined;
-
-export function buildModuleGraph(
-  fileToModulesMap: ViteDevServer['moduleGraph']['fileToModulesMap']
-): ModuleGraph {
-  const moduleGraph: ModuleGraph = new Map();
-  const moduleNodeMap = new WeakMap<object, ModuleNode>();
-
-  const getOrCreateModuleNode = (
-    viteModuleNode: {
-      file: string | null;
-      type: ViteModuleNode['type'];
-      importers: Set<ViteModuleNode>;
-      importedModules: Set<ViteModuleNode>;
-    },
-    fallbackFile?: string
-  ): ModuleNode | undefined => {
-    const file = viteModuleNode.file ?? fallbackFile;
-    if (!file) {
-      return undefined;
-    }
-
-    const existingNode = moduleNodeMap.get(viteModuleNode);
-    if (existingNode) {
-      return existingNode;
-    }
-
-    const moduleNode: ModuleNode = {
-      file,
-      type: viteModuleNode.type,
-      importers: new Set(),
-      importedModules: new Set(),
-    };
-    moduleNodeMap.set(viteModuleNode, moduleNode);
-
-    const moduleSet = moduleGraph.get(file) ?? new Set<ModuleNode>();
-    moduleSet.add(moduleNode);
-    moduleGraph.set(file, moduleSet);
-
-    return moduleNode;
-  };
-
-  fileToModulesMap.forEach((viteModuleSet, filePath) => {
-    viteModuleSet.forEach((viteModuleNode) => {
-      const moduleNode = getOrCreateModuleNode(viteModuleNode, filePath);
-      if (moduleNode) {
-        viteModuleNode.importers.forEach((importer) => {
-          const importerNode = getOrCreateModuleNode(importer);
-          if (importerNode) {
-            moduleNode.importers.add(importerNode);
-          }
-        });
-        viteModuleNode.importedModules.forEach((importedModule) => {
-          const importedModuleNode = getOrCreateModuleNode(importedModule);
-          if (importedModuleNode) {
-            moduleNode.importedModules.add(importedModuleNode);
-          }
-        });
-      }
-    });
-  });
-
-  return moduleGraph;
-}
 
 function notifyListeners(moduleGraph: ModuleGraph): void {
   listeners.forEach((listener) => {
